requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.5
fastapi==0.129.0
google-auth==2.48.0
google-auth-oauthlib==1.2.4
gspread==6.2.1
h11==0.16.0
idna==3.11
Jinja2==3.1.6
MarkupSafe==3.0.3
numpy==2.4.2
oauthlib==3.3.1
pandas==3.0.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.41.0
